IntelliPet/requirements.txt
adafruit-circuitpython-servokit==1.3.10
SpeechRecognition==3.8.1
pyaudio==0.2.13
RPi.GPIO==0.7.1
